Fix build file

+++ build.ps1
@@ -1,20 +1,54 @@
-$projName = "uCommunity.Core";
-$configuration = "Release";
-$tagPrefix = "Core";
+# $projName = "uCommunity.Core";
+# $configuration = "Release";
+# $tagPrefix = "Core";
 
+param 
+( 
+    [string]$projName = "",
+    [string]$configuration = "",
+    [string]$tagPrefix = ""
+)
+
+if([string]::IsNullOrEmpty($projName)){
+    throw [System.ArgumentException] "Invalid project name!"
+}
+
+if([string]::IsNullOrEmpty($configuration)){
+    throw [System.ArgumentException] "Invalid configuration name!"
+}
+
+if([string]::IsNullOrEmpty($tagPrefix)){
+    throw [System.ArgumentException] "Invalid tag prefix!"
+}
 
 $nuget = ".\src\.nuget\nuget";
+
+if(-not (Test-Path $nuget)){
+    throw [System.IO.FileNotFoundException] "Can't find nuget.exe. Should be in $nuget";
+}
+
+$nugetConfig = ".\src\.nuget\nuget.config";
+
+if(-not (Test-Path $nugetConfig)){
+    throw [System.IO.FileNotFoundException] "Can't find nuget.config. Should be in $nugetConfig";
+}
+
 $projFile = Get-ChildItem -Recurse "$projName.csproj";
 
+if(-not (Test-Path $projFile)){
+     throw [System.IO.FileNotFoundException] "Could not find project file - '$projName'";
+}
 
-Invoke-Expression "$nuget pack -Build $projFile -Properties Configuration=$configuration";
+Invoke-Expression "$nuget restore '$projFile' -ConfigFile '$nugetConfig'";
+
+Invoke-Expression "$nuget pack -Build '$projFile' -Properties Configuration=$configuration";
 
 $nugetFile = Get-ChildItem "$projName*.nupkg";
 
 if (Test-Path $nugetFile) {
     $nugetsFolder = "C:\inetpub\Nuget\Packages";
 
-    Invoke-Expression "$nuget push $nugetFile -ApiKey !QA2ws3ed -Source http://nuget.compent.dk/ ";
+    Invoke-Expression "$nuget push '$nugetFile' -ApiKey !QA2ws3ed -Source http://nuget.compent.dk/ ";
 
     Remove-Item $nugetFile;
 
@@ -29,6 +63,3 @@ if (Test-Path $nugetFile) {
      Unblock-File $assemblyPath
      Unblock-File $projFile
 }
-
-
-
